+++ .vscode/settings.json
@@ -22,10 +22,10 @@
     "wsgi",
     "wtforms"
   ],
-  "python.testing.unittestArgs": ["-v", "-s", "./tests", "-p", "*test.py"],
+  // "python.testing.unittestArgs": ["-v", "-s", "./tests", "-p", "*test.py"],
   "python.testing.pytestEnabled": true,
   "python.testing.unittestEnabled": false,
-  "python.testing.pytestArgs": ["tests"],
+  // "python.testing.pytestArgs": ["tests"],
   "[python]": {
     "editor.defaultFormatter": "ms-python.black-formatter"
   }
