+++ app/commands/__init__.py
@@ -20,7 +20,7 @@ def init(app: Flask):
         @click.option("--count", default=100, type=int)
         def db_populate(count: int):
             """Fill DB by dummy data."""
-            from tests.db import populate
+            from test_flask.db import populate
 
             populate(count)
             print(f"DB populated by {count} instancies")
