+++ test_api/conftest.py
@@ -1,0 +1,3 @@
+from dotenv import load_dotenv
+
+load_dotenv("test_api/test.env")
