+++ test_api/test.env
@@ -1,0 +1,1 @@
+IS_API=true
