+++ test_api/test_users.py
@@ -1,0 +1,5 @@
+def test_get_me():
+    from config import config
+
+    CFG = config("testing")
+    assert CFG.IS_API
